+++ src/osm_easy_api/api/endpoints/gpx.py
@@ -1,9 +1,12 @@
 import shutil
 
 from typing import TYPE_CHECKING
+from xml.dom import minidom
+
 if TYPE_CHECKING:
     from ...api import Api
 
+from ...data_classes import GpxFile, Visibility
 # TODO: GPX full support and parser
 
 class Gpx_Container:
@@ -25,13 +28,13 @@ class Gpx_Container:
         with open(file_to, "wb") as f_to:
             shutil.copyfileobj(response.raw, f_to)
 
-    def create(self, file_from: str, description: str, visibility: str, tags: list[str] | None = None) -> int:
+    def create(self, file_from: str, description: str, visibility: Visibility, tags: list[str] | None = None) -> int:
         """Uploads a GPX file or archive of GPX files.
 
         Args:
             file_from (str): Path to file to be uploaded.
             description (str): The trace description.
-            visibility (str): 'private', 'public', 'trackable' or 'identifiable'. See https://wiki.openstreetmap.org/wiki/Visibility_of_GPS_traces for more info.
+            visibility (Visibility): See https://wiki.openstreetmap.org/wiki/Visibility_of_GPS_traces for more info.
             tags (Tags | None, optional): Tags for the trace. Defaults to None.
 
         Returns:
@@ -48,8 +51,22 @@ class Gpx_Container:
                 "file": f,
                 "description": (None, description),
                 "tags": (None, tags_string),
-                "visibility": (None, visibility)
+                "visibility": (None, visibility.value)
             }
 
-            response = self.outer._request(self.outer._RequestMethods.POST, url=self.outer._url.gpx["create"], files=files)
+            response = self.outer._request(method=self.outer._RequestMethods.POST, url=self.outer._url.gpx["create"], files=files)
             return int(response.text)
+        
+    def update(self, gpx_file: GpxFile) -> None:
+        """Updates a GPX file.
+
+        Args:
+            gpx_file (GpxFile): GPX file with a new description and/or tags.
+        """
+        root = minidom.Document()
+        xml = root.createElement("osm")
+        root.appendChild(xml)
+        xml.appendChild(gpx_file._to_xml())
+        xml_str = root.toprettyxml()
+        
+        self.outer._request(method=self.outer._RequestMethods.PUT, url=self.outer._url.gpx["update"].format(id=gpx_file.id), body=xml_str)