+++ src/osm_easy_api/data_classes/__init__.py
@@ -9,3 +9,5 @@ from .tags import Tags
 from .changeset import Changeset
 from .user import User
 from .note import Note, Comment
+
+from .GpxFile import GpxFile, Visibility